+++ api-health-classification/database/migrations/2019_09_15_210424_create_table_balita_training.php
@@ -16,8 +16,8 @@ class CreateTableBalitaTraining extends Migration
         Schema::create('balita_training', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->integer('umur');
-            $table->integer('berat_badan');
-            $table->integer('tinggi_badan');
+            $table->double('berat_badan');
+            $table->double('tinggi_badan');
             $table->string('status');
             $table->timestamps();
         });
