+++ api-health-classification/database/migrations/2019_09_15_210434_create_table_balita_test.php
@@ -16,8 +16,8 @@ class CreateTableBalitaTest extends Migration
         Schema::create('balita_test', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->integer('umur');
-            $table->integer('berat_badan');
-            $table->integer('tinggi_badan');
+            $table->double('berat_badan');
+            $table->double('tinggi_badan');
             $table->string('status');
             $table->timestamps();
         });
