+++ api-health-classification/database/migrations/2019_09_15_210455_create_table_bumil_classification.php
@@ -20,8 +20,8 @@ class CreateTableBumilClassification extends Migration
           $table->string('nama');
           $table->integer('usia_bumil');
           $table->integer('usia_kehamilan');
-          $table->integer('berat_badan');
-          $table->integer('tinggi_badan');
+          $table->double('berat_badan');
+          $table->double('tinggi_badan');
           $table->string('LILA');
           $table->string('KEK');
           $table->string('status');
